+++ src/TabularInput.php
@@ -21,8 +21,9 @@ use unclead\widgets\renderers\TableRenderer;
  */
 class TabularInput extends Widget
 {
-    const POS_HEADER    = 0;
-    const POS_ROW       = 1;
+    const POS_HEADER    = TableRenderer::POS_HEADER;
+    const POS_ROW       = TableRenderer::POS_ROW;
+    const POS_FOOTER    = TableRenderer::POS_FOOTER;
 
     /**
      * @var array